Added sections for BlipTV and Vimeo and shortened code
Added BlipTV and Vimeo sections and types, so videos from these sites
can be added. Vimeo actually uses RSS feed, but added section and
ability to just add Vimeo url to prevent confusion. Also created a
function for a couple areas of code that were repeated often to shorten
the length of the code.